+++ R/utils_helpers.R
@@ -1,11 +1,15 @@
 regions <- function() {
-  unique(dpc_covid19_ita_regioni$denominazione_regione)
+  sort(unique(dpc_covid19_ita_regioni$denominazione_regione))
+}
+
+provinces <- function() {
+  sort(unique(dpc_covid19_ita_province$denominazione_provincia))
 }
 
 measures <- function(level = c("national", "regional", "provincial")) {
   level <- match.arg(level)
 
-  switch(level,
+  sort(switch(level,
     national = ,
     regional = c(
       "ricoverati_con_sintomi", "terapia_intensiva",
@@ -13,5 +17,5 @@ measures <- function(level = c("national", "regional", "provincial")) {
       "totale_attualmente_positivi", "dimessi_guariti",
       "deceduti", "totale_casi", "tamponi"),
     provincial = "totale_casi"
-  )
+  ))
 }