+++ NEWS.md
@@ -1,3 +1,4 @@
+* Added provincial module
 * National module converted in italian
 * Menu and title in italian